Add protection for first sector and override of protection on config bytes, add code security bits
Maybe there is memory above 0x800 that should be protected in a similar
way but my addition only focuses on the first sector and I think uses a
method making a small price to pay to gaurd against careless programming
and unnecessary angst for beginners.

I add the lock and unlock methods I have chosen for myself here because
it seems the logical place to put them.